Fixed Prefab scaling issues
Adjusted a bunch of prefabs to structure them consistently with a Model child object that can be arbitrarily scaled while leaving the main object uniformly scaled to obey rules. This fixes a lot of the issues I was seeing before but I haven't extensively tested edge cases with it yet.