MOD: Added Content field to HMT Response Format
Added a content field to be separate from the message field in responses to the HMT Interface calls. This way message is guaranteed to be a string, while content is generally meant to be a JSON encoded blob but could also be a string. get_state now returns the state in "content" instead of "message". All other commands are currently unchanged.