REFACTOR: unified NetworkMiddleware Function Signatures
Unified all NetworkMiddleware function signatures to start with the character ID then follow with necessary parameters. Also naming now follows a pattern of "CallX" being the public function that either makes an RPC or calls the local version, and "XLocal" being the private function that was the Rpc. In the process I also wired up the HMTInterface