REFACTOR/FIX: Redefined Level Specs
Added Level names and dimensions to level spec files so that they can be more easily read and accessed in logging and HMT interface functions. Also fixed the logical transposition issue between the state in game and how it was presented to the HMT Interface. Note that maps will appear to be flipped around the "up" axis because they were previously being generated wrong.