REFACTOR: Updated UIManager Tooltip Overlay
Made several changes to the UIManager's system for the challenge tooltip. Renamed and rewired all of the UI element references so they don't have to call GetComponent. Also streamlined the processing to make it easier to send a unified log message and gated the update logic behind a target reference to avoid duplicate logging messages.