FIX: HMT Interface will now wait for initialization
Originally, the HMTItnterface would response with an 1002 error if an agent tried to take an action before the game was ready. Not it will just wait for the scene to be properly initialized and respond accordingly. This is because the GetReady state is used both for first time initialization and when transitioning levels. Previously, it assumed only the former was true.