ADD: added instantMode flags for all AI games
you can now set the executionStepTime for compiled versions of the game by providing a -stepTime command line arg. Alternatively you can provide a single flag of -instantMode, which sets the step time to 0. All aspects of the game logic and UI animation are now based on executionStepTime as well. In principle the game should run nearly instantly when the time is set to 0.